requirements.txt
openai>=1.0.0
tenacity>=8.0.0
python-dotenv>=1.0.0
google-genai>=0.3.0
nltk>=3.8.1
datasets>=2.14.0
pandas>=2.0.0
PyYAML>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
